+++ ClientApp/package.json
@@ -46,6 +46,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.0.0",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "~4.9.5"
   },
   "overrides": {
