+++ ClientApp/src/app/components/product/product.module.ts
@@ -2,12 +2,13 @@ import { NgModule } from '@angular/core';
 import {CommonModule, NgOptimizedImage} from '@angular/common';
 import { NgxMaskModule } from 'ngx-mask';
 
-import { ProductListComponent } from "@components/product/list";
+import { ProductListComponent } from "@components/product/list/product-list.component";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import { ProductCardComponent } from '@components/product/card';
-import {RegisterProductComponent} from "@components/product/register";
-import {SharedModule} from "../shared/shared.module";
-import {ProductService} from "@service/product";
+import {RegisterProductComponent} from "@components/product/register/register-product.component";
+
+import { ProductCardComponent } from '@components/product/product-card/product-card.component';
+import {SharedModule} from "@shared/shared.module";
+import {ProductService} from "./product.service";
 
 @NgModule({
   declarations: [
