+++ ClientApp/src/app/components/product/register/register-product.component.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import {ProductService} from "@service/product";
+import {ProductService} from "@components/product/product.service";
 @Component({
   selector: 'app-register-product',
   templateUrl: './register-product.component.html',
