+++ ClientApp/src/app/components/shared/shared.module.ts
@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {HeaderComponent} from "./header/header.component";
+import {HeaderComponent} from "@components/shared/header/header.component";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import { RouterLink } from "@angular/router";
-import {SideNavComponent} from "@components/shared/sidenav";
-
-
+import {SideNavComponent} from "@shared/side-nav/side-nav.component";
 
 @NgModule({
   declarations: [
