+++ ClientApp/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compileOnSave": false,
   "compilerOptions": {
-    "baseUrl": "./",
+    //"baseUrl": "./src",
     "outDir": "./dist/out-tsc",
     "forceConsistentCasingInFileNames": true,
     "strict": true,
@@ -20,25 +20,28 @@
       "es2018",
       "dom"
     ],
-    "useDefineForClassFields": false,
     "paths": {
-      "@components/shared/sidenav": ["src/app/components/shared/side-nav/side-nav.component.ts"],
-      "@components/shared/header": ["src/app/components/shared/header/header.component.ts"],
-      "@components/product/card": ["src/app/components/product/product-card/product-card.component.ts"],
-      "@components/product/list": ["src/app/components/product/list/product-list.component.ts"],
-      "@components/product/module": ["src/app/components/product/product.module.ts"],
-      "@components/product/register": ["src/app/components/product/register/register-product.component.ts"],
-      "@components/home": ["src/app/components/home/home.component.ts"],
-      "@enums/screen-dimensions": ["src/app/enums/screen-dimensions.ts"],
-      "@interfaces/product": ["src/app/interfaces/product.ts"],
-      "@service/product": ["src/app/components/product/product.service.ts"]
-    }
+      "@shared/*": [
+        "./src/app/components/shared/*",
+      ],
+      "@components/*": [
+        "./src/app/components/*"
+      ],
+      "@interfaces/*": [
+        "./src/app/interfaces/*"
+      ],
+      "@enums/*": [
+        "./src/app/enums/*"
+      ],
+    },
+    "useDefineForClassFields": false,
+
   },
   "angularCompilerOptions": {
     "enableI18nLegacyMessageIdFormat": false,
     "strictInjectionParameters": true,
     "strictInputAccessModifiers": true,
-    "strictTemplates": true
+    "strictTemplates": true,
   },
   "files": [
     "app.module.ts"
